# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vatel AI REST API",
    "description": "Programmatic access for agents, telephony, and session tokens. Send `Authorization: Bearer <organization API key>` on all requests unless noted. Most error responses return JSON with an `error` string message.",
    "version": "0.2.1"
  },
  "servers": [
    {
      "url": "https://api.vatel.ai",
      "description": "Production"
    }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/v1/sip-trunks": {
      "get": {
        "tags": ["sip-trunks"],
        "summary": "List SIP trunks",
        "operationId": "listSipTrunks",
        "responses": {
          "200": {
            "description": "SIP trunks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/SipTrunk" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["sip-trunks"],
        "summary": "Create SIP trunk",
        "operationId": "createSipTrunk",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SipTrunkCreateInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created trunk",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SipTrunk" }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "409": {
            "description": "Inbound SIP username already in use",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "503": {
            "description": "SIP connectivity is not available in this environment",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/v1/sip-trunks/{id}/assignments": {
      "get": {
        "tags": ["sip-trunks"],
        "summary": "List agent assignments on SIP trunk",
        "operationId": "listSipTrunkAgentAssignments",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "SIP trunk id",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Assignments: `number` for digest-auth trunks, `alternate_number` for ACL trunks (globally unique).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/SipTrunkAgentAssignment" }
                }
              }
            }
          },
          "400": { "description": "Invalid trunk id" },
          "401": { "description": "Unauthorized" },
          "404": {
            "description": "SIP trunk not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["sip-trunks"],
        "summary": "Create agent assignment on SIP trunk",
        "description": "Maps an org agent to this trunk. Use `number` when inbound auth is digest (unique on this trunk). Use `alternate_number` when inbound auth is ACL (globally unique alternate dial target). Same rules as the dashboard integrations UI.",
        "operationId": "createSipTrunkAgentAssignment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "SIP trunk id",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SipTrunkAgentAssignmentCreateInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created assignment",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SipTrunkAgentAssignment" }
              }
            }
          },
          "400": {
            "description": "Validation error or agent not in organization",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": {
            "description": "SIP trunk or agent not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "409": {
            "description": "Number or alternate number already in use",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/v1/sip-trunks/assignments/{assignmentId}": {
      "get": {
        "tags": ["sip-trunks"],
        "summary": "Get SIP trunk assignment",
        "operationId": "getSipTrunkAgentAssignment",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "required": true,
            "description": "Assignment id",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Assignment",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SipTrunkAgentAssignment" }
              }
            }
          },
          "400": { "description": "Invalid assignment id" },
          "401": { "description": "Unauthorized" },
          "404": {
            "description": "Assignment not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["sip-trunks"],
        "summary": "Update SIP trunk assignment",
        "description": "Patch `number` and/or `alternate_number`. Send `null` to clear a field. Omitted keys are left unchanged. Assignment id is globally unique.",
        "operationId": "patchSipTrunkAgentAssignment",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "required": true,
            "description": "Assignment id",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SipTrunkAgentAssignmentPatchInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated assignment",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SipTrunkAgentAssignment" }
              }
            }
          },
          "400": {
            "description": "Invalid id or body",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": {
            "description": "Assignment not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "409": {
            "description": "Number or alternate number already in use",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["sip-trunks"],
        "summary": "Delete SIP trunk assignment",
        "description": "Assignment id is globally unique.",
        "operationId": "deleteSipTrunkAgentAssignment",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "required": true,
            "description": "Assignment id",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "204": { "description": "Deleted" },
          "400": { "description": "Invalid assignment id" },
          "401": { "description": "Unauthorized" },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/v1/sip-trunks/{id}": {
      "get": {
        "tags": ["sip-trunks"],
        "summary": "Get SIP trunk",
        "operationId": "getSipTrunk",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "SIP trunk id",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "SIP trunk",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SipTrunk" }
              }
            }
          },
          "400": { "description": "Invalid id" },
          "401": { "description": "Unauthorized" },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["sip-trunks"],
        "summary": "Update SIP trunk",
        "operationId": "patchSipTrunk",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "SIP trunk id",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SipTrunkUpdateInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated trunk",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SipTrunk" }
              }
            }
          },
          "400": {
            "description": "Validation error or no fields to update",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "409": {
            "description": "Inbound SIP username already in use",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["sip-trunks"],
        "summary": "Delete SIP trunk",
        "operationId": "deleteSipTrunk",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "SIP trunk id",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "204": { "description": "Deleted" },
          "400": { "description": "Invalid id" },
          "401": { "description": "Unauthorized" },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "Send your organization API key as `Authorization: Bearer <key>`."
      }
    },
    "schemas": {
...
      "CallerTransformType": {
        "type": "string",
        "description": "How to adjust caller id before routing.",
        "enum": [
          "add_prefix",
          "add_suffix",
          "replace",
          "strip_digits_end",
          "strip_digits_start"
        ]
      },
      "SIPTrunkAuthType": {
        "type": "string",
        "description": "How inbound SIP is authenticated. `acl`: carrier traffic must match the ACL; `inbound_host` is the ACL-validated host. `digest`: SIP digest authentication. `none`: no inbound auth configured.",
        "enum": ["digest", "acl", "none"]
      },
      "RegistrationStatus": {
        "type": "string",
        "description": "Last known SIP registration outcome for the trunk leg.",
        "enum": ["not_registered", "pending", "registered", "failed", "auth_failed"]
      },
      "SipTrunkCallerIDTransform": {
        "type": "object",
        "description": "Rule applied to caller id on this trunk.",
        "required": ["type"],
        "properties": {
          "type": { "$ref": "#/components/schemas/CallerTransformType" },
          "value": { "type": "string", "description": "Primary operand (e.g. prefix string)." },
          "value2": { "type": "string", "description": "Secondary operand when the transform requires it." },
          "number": { "type": "integer", "description": "Numeric parameter (e.g. digit count) when applicable." }
        }
      },
      "SipTrunk": {
        "type": "object",
        "description": "SIP connectivity profile for a supported PBX integration.",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "created_at": { "type": "string", "format": "date-time" },
          "pbx": {
            "type": "string",
            "description": "PBX vendor preset (`3cx`, `yeastar`, `webex`) applying default trunk behavior, or `generic` for other / custom PBX setups where you supply hosts and credentials explicitly.",
            "enum": ["3cx", "yeastar", "webex", "generic"]
          },
          "caller_id_transforms": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SipTrunkCallerIDTransform" },
            "description": "Ordered transforms applied to caller id."
          },
          "inbound_host": {
            "type": "string",
            "description": "Inbound SIP host or domain. When `inbound_auth_type` is `acl`, this is the host validated against the ACL (allowed source for inbound calls)."
          },
          "inbound_auth_type": { "$ref": "#/components/schemas/SIPTrunkAuthType" },
          "inbound_sip_username": { "type": "string", "description": "SIP user for inbound registration if used." },
          "inbound_registration_status": { "$ref": "#/components/schemas/RegistrationStatus" },
          "outbound_host": {
            "type": "string",
            "description": "SIP host used for outbound legs: platform-initiated calls, on-call flows, and transfers dial out via this host (your carrier or SBC)."
          },
          "outbound_sip_username": { "type": "string", "description": "Credentials username for outbound, if any." },
          "register": { "type": "boolean", "description": "Whether the platform registers this trunk with the carrier." },
          "outbound_registration_status": { "$ref": "#/components/schemas/RegistrationStatus" },
          "remain_in_dialog": { "type": "boolean", "description": "Keeping SIP dialog routing behavior when true." }
        }
      },
      "SipTrunkCreateInput": {
        "type": "object",
        "description": "Create a SIP trunk; passwords are write-only on the wire.",
        "required": ["pbx"],
        "properties": {
          "pbx": {
            "type": "string",
            "description": "PBX vendor preset or `generic` for other/custom PBX configurations.",
            "enum": ["3cx", "yeastar", "webex", "generic"]
          },
          "caller_id_transforms": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SipTrunkCallerIDTransform" }
          },
          "inbound_host": {
            "type": "string",
            "nullable": true,
            "description": "Inbound SIP host or domain; when inbound auth is `acl`, this host is ACL-validated for inbound traffic."
          },
          "inbound_auth_type": { "$ref": "#/components/schemas/SIPTrunkAuthType", "nullable": true },
          "inbound_sip_username": { "type": "string", "nullable": true },
          "inbound_sip_password": { "type": "string", "nullable": true, "description": "Inbound SIP secret if required." },
          "outbound_host": {
            "type": "string",
            "nullable": true,
            "description": "Host to dial for outbound calls, on-call, and transfers."
          },
          "outbound_sip_username": { "type": "string", "nullable": true },
          "outbound_sip_password": { "type": "string", "nullable": true, "description": "Outbound SIP secret if required." },
          "register": { "type": "boolean", "nullable": true, "description": "Enable outbound registration when true." },
          "remain_in_dialog": { "type": "boolean", "nullable": true }
        }
      },
      "SipTrunkUpdateInput": {
        "type": "object",
        "description": "Patch only fields you need; empty string clears optional text fields per server rules.",
        "properties": {
          "pbx": {
            "type": "string",
            "enum": ["3cx", "yeastar", "webex", "generic"],
            "description": "Vendor preset or `generic` for other/custom PBX."
          },
          "caller_id_transforms": {
            "type": "array",
            "nullable": true,
            "items": { "$ref": "#/components/schemas/SipTrunkCallerIDTransform" },
            "description": "Replace transforms; empty array clears."
          },
          "inbound_host": {
            "type": "string",
            "nullable": true,
            "description": "Inbound host; with `acl` auth, must match the ACL-validated host."
          },
          "inbound_auth_type": { "$ref": "#/components/schemas/SIPTrunkAuthType", "nullable": true },
          "inbound_sip_username": { "type": "string", "nullable": true },
          "inbound_sip_password": { "type": "string", "nullable": true, "description": "Set to rotate inbound secret." },
          "outbound_host": {
            "type": "string",
            "nullable": true,
            "description": "Outbound dial host for calls, on-call, and transfers."
          },
          "outbound_sip_username": { "type": "string", "nullable": true },
          "outbound_sip_password": { "type": "string", "nullable": true },
          "register": { "type": "boolean", "nullable": true },
          "remain_in_dialog": { "type": "boolean", "nullable": true }
        }
      },
      "SipTrunkAgentAssignment": {
        "type": "object",
        "description": "Binds an org agent to a SIP trunk. With digest inbound auth, routing uses `number`. With ACL inbound auth, routing uses `alternate_number` (globally unique).",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "agent_id": { "type": "string", "format": "uuid" },
          "sip_trunk_id": { "type": "string", "format": "uuid" },
          "number": {
            "type": "string",
            "description": "Number to dial for this agent when the trunk uses digest inbound auth; must be unique among assignments on this trunk."
          },
          "alternate_number": {
            "type": "string",
            "description": "Globally unique alternate dial string when the trunk uses ACL inbound auth; used instead of per-trunk `number` routing in that mode."
          },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "SipTrunkAgentAssignmentCreateInput": {
        "type": "object",
        "description": "Create assignment. Use `number` for digest-auth trunks (unique on this trunk). Use `alternate_number` for ACL trunks (globally unique). Both may be set if you use both routing styles.",
        "required": ["agent_id"],
        "properties": {
          "agent_id": { "type": "string", "format": "uuid" },
          "number": {
            "type": "string",
            "description": "Target number to dial for digest inbound auth; unique among assignments on this SIP trunk."
          },
          "alternate_number": {
            "type": "string",
            "description": "Globally unique alternate number to dial when the trunk uses ACL inbound auth."
          }
        }
      },
      "SipTrunkAgentAssignmentPatchInput": {
        "type": "object",
        "description": "Patch only fields present in the JSON. Use null to clear `number` or `alternate_number`.",
        "properties": {
          "number": {
            "type": "string",
            "nullable": true,
            "description": "Digest-auth routing number (unique per trunk); null clears."
          },
          "alternate_number": {
            "type": "string",
            "nullable": true,
            "description": "ACL routing alternate dial string (globally unique); null clears."
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error body for most non-2xx JSON responses on this API.",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "description": "Human-readable message; may include validation detail." }
        }
      }
    }
  }
}